requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.15
botocore==1.31.15
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
Django==4.2.3
django-environ==0.10.0
django-filter==23.2
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
exceptiongroup==1.1.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
s3transfer==0.6.1
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
whitenoise==6.5.0
